# src/main/java/com/dmanso/pizzashackSpringboot/model/MenuItem.java
package com.dmanso.pizzashackSpringboot.model;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Entity
@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class MenuItem {


	@Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    @Column(updatable = false, nullable = false)
	private String menuId;

	@Column
	private String name;

	@Column
	private String description;

	@Column
	private String price;
	
	@Column
	private String icon;	
		
}